requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dotenv==0.9.9
et_xmlfile==2.0.0
google-auth==2.40.3
google-genai==1.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
json_repair==0.49.0
markdownify==0.13.1
numpy==1.26.3
openpyxl==3.1.5
pandas==2.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jobspy==1.1.82
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tenacity==9.1.2
tls-client==1.0.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
